+++ Example/Pods/Target Support Files/Pods-XMLMapper_Example/Pods-XMLMapper_Example-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Pods_XMLMapper_Example : NSObject
-@end
-@implementation PodsDummy_Pods_XMLMapper_Example
-@end