+++ XMLMapper/Classes/Requests/SOAPInformation.swift
@@ -11,16 +11,14 @@ open class SOAPInformation: XMLMappable {
     public var nodeName: String!
     
     private var xmlnsInformation: String?
-    internal(set) var informationName: String?
+    var informationName: String?
     
     public init(informationName: String, nameSpace: String) {
         self.informationName = informationName
         self.xmlnsInformation = nameSpace
     }
     
-    required public init(map: XMLMap) {
-        
-    }
+    required public init?(map: XMLMap) {}
     
     open func mapping(map: XMLMap) {
         xmlnsInformation <- map.attributes["xmlns:m"]