+++ XMLMapper/Classes/Requests/SOAPMessage.swift
@@ -11,16 +11,14 @@ open class SOAPMessage: XMLMappable {
     public var nodeName: String!
     
     private var xmlnsMessage: String?
-    internal(set) var soapAction: String?
+    var soapAction: String?
     
     public init(soapAction: String, nameSpace: String) {
         self.soapAction = soapAction
         self.xmlnsMessage = nameSpace
     }
     
-    required public init(map: XMLMap) {
-        
-    }
+    required public init?(map: XMLMap) {}
     
     open func mapping(map: XMLMap) {
         xmlnsMessage <- map.attributes["xmlns:m"]