+++ XMLMapper/Classes/XMLObjectParser.swift
@@ -31,6 +31,7 @@ class XMLObjectParser: NSObject {
     fileprivate var preserveComments: Bool
     fileprivate var wrapRootNode: Bool
     fileprivate var cdataAsString: Bool
+    fileprivate var keepNodesOrder: Bool
     fileprivate var attributesMode: XMLObjectParserAttributesMode = .prefixed
     fileprivate var nodeNameMode: XMLObjectParserNodeNameMode = .always
     
@@ -44,12 +45,13 @@ class XMLObjectParser: NSObject {
     
     internal required override init() {
         collapseTextNodes = true
-        stripEmptyNodes = true
+        stripEmptyNodes = false
         trimWhiteSpace = true
         alwaysUseArrays = false
         preserveComments = false
         wrapRootNode = false
         cdataAsString = false
+        keepNodesOrder = true
     }
     
     // MARK: - Basic interface function
@@ -77,6 +79,7 @@ class XMLObjectParser: NSObject {
         preserveComments = options.contains(.preserveComments)
         wrapRootNode = options.contains(.wrapRootNode)
         cdataAsString = options.contains(.cdataAsString)
+        keepNodesOrder = options.contains(.keepNodesOrder)
 
         if options.contains(.prefixedAttributes) {
             attributesMode = .prefixed
@@ -193,10 +196,19 @@ extension XMLObjectParser: XMLParserDelegate {
                 existingArray.add(node)
             } else if let existing = existing {
                 top?[elementName] = NSMutableArray(array: [existing, node])
-            } else if alwaysUseArrays {
-                top?[elementName] = NSMutableArray(object: node)
             } else {
-                top?[elementName] = node
+                if alwaysUseArrays {
+                    top?[elementName] = NSMutableArray(object: node)
+                } else {
+                    top?[elementName] = node
+                }
+                if keepNodesOrder {
+                    if let nodesOrder = top?[XMLParserConstant.Key.nodesOrder] as? NSMutableArray {
+                        nodesOrder.add(elementName)
+                    } else {
+                        top?[XMLParserConstant.Key.nodesOrder] = NSMutableArray(object: elementName)
+                    }
+                }
             }
             stack?.append(node)
         }