+++ XMLMapper/Classes/XMLSerialization+ReadingOptions.swift
@@ -21,24 +21,25 @@ extension XMLSerialization {
         public static let alwaysUseArrays = ReadingOptions(rawValue: 1 << 3)
         public static let preserveComments = ReadingOptions(rawValue: 1 << 4)
         public static let wrapRootNode = ReadingOptions(rawValue: 1 << 5)
+        public static let keepNodesOrder = ReadingOptions(rawValue: 1 << 6)
         
-        public static let prefixedAttributes = ReadingOptions(rawValue: 1 << 6)
-        public static let unprefixedAttributes = ReadingOptions(rawValue: 1 << 7)
-        public static let dictionaryAttributes = ReadingOptions(rawValue: 1 << 8)
-        public static let discardAttributes = ReadingOptions(rawValue: 1 << 9)
+        public static let prefixedAttributes = ReadingOptions(rawValue: 1 << 7)
+        public static let unprefixedAttributes = ReadingOptions(rawValue: 1 << 8)
+        public static let dictionaryAttributes = ReadingOptions(rawValue: 1 << 9)
+        public static let discardAttributes = ReadingOptions(rawValue: 1 << 10)
         
-        public static let rootOnlyNodeName = ReadingOptions(rawValue: 1 << 10)
-        public static let alwaysNodeName = ReadingOptions(rawValue: 1 << 11)
-        public static let neverNodeName = ReadingOptions(rawValue: 1 << 12)
+        public static let rootOnlyNodeName = ReadingOptions(rawValue: 1 << 11)
+        public static let alwaysNodeName = ReadingOptions(rawValue: 1 << 12)
+        public static let neverNodeName = ReadingOptions(rawValue: 1 << 13)
         
         public static let cdataAsString = ReadingOptions(rawValue: 1 << 13)
         
         public static let `default`: ReadingOptions = [
             .collapseTextNodes,
-            .stripEmptyNodes,
             .trimWhiteSpace,
+            .keepNodesOrder,
             .prefixedAttributes,
-            .alwaysNodeName
+            .alwaysNodeName,
         ]
     }
 }