+++ XMLMapperTests/Tests/XMLDataTransformTests.swift
@@ -40,7 +40,7 @@ class DataType: XMLMappable {
 		
 	}
 	
-	required init(map: XMLMap){
+	required init?(map: XMLMap){
 		
 	}